Move candidate gen
