+++ src/theory/quantifiers/ematching/candidate_generator.cpp
@@ -12,7 +12,7 @@
  ** \brief Implementation of theory uf candidate generator class
  **/
 
-#include "theory/quantifiers/candidate_generator.h"
+#include "theory/quantifiers/ematching/candidate_generator.h"
 #include "options/quantifiers_options.h"
 #include "theory/quantifiers/inst_match.h"
 #include "theory/quantifiers/instantiate.h"
