+++ src/theory/quantifiers/ematching/candidate_generator.h
